+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## [ 0.1.1 ] - 2022-6-22 15:48:24
+
+> BENS-0002 Simplify Release Workflow
+
+### Changed
+
+- `d98b829` - BENS-0002 added workflow + install script
+
+
+
 ## [ 0.1.0 ] - 2022-6-21 21:51:9
 
 > BENS-0001 Backup Script
@@ -16,5 +26,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `3cee613` - BENS-0001 ensured that latest snapshot will never be deleted
 - `58f1865` - BENS-0001 implemented snapshot code with backup rotation
-